+++ eras/conway/impl/src/Cardano/Ledger/Conway/Rules/Ratify.hs
@@ -42,8 +42,8 @@ import Cardano.Ledger.Conway.Governance (
 import Cardano.Ledger.Conway.Governance.Procedures (committeeMembersL)
 import Cardano.Ledger.Conway.PParams (
   ConwayEraPParams,
-  ppCommitteeTermLimitL,
-  ppMinCommitteeSizeL,
+  ppCommitteeMaxTermLengthL,
+  ppCommitteeMinSizeL,
  )
 import Cardano.Ledger.Conway.Rules.Enact (EnactSignal (..), EnactState (..))
 import Cardano.Ledger.Credential (Credential (..))
@@ -299,9 +299,9 @@ prevActionAsExpected _ _ = True -- for the other actions, the previous action is
 
 validCommitteeSize :: ConwayEraPParams era => StrictMaybe (Committee era) -> PParams era -> Bool
 validCommitteeSize committee pp =
-  let minCommitteeSize = pp ^. ppMinCommitteeSizeL
+  let committeeMinSize = pp ^. ppCommitteeMinSizeL
       members = foldMap' (^. committeeMembersL) committee
-   in Map.size members >= fromIntegral minCommitteeSize
+   in Map.size members >= fromIntegral committeeMinSize
 
 validCommitteeTerm ::
   ConwayEraPParams era =>
@@ -310,7 +310,7 @@ validCommitteeTerm ::
   EpochNo ->
   Bool
 validCommitteeTerm committee pp currentEpoch =
-  let maxCommitteeTerm = pp ^. ppCommitteeTermLimitL
+  let maxCommitteeTerm = pp ^. ppCommitteeMaxTermLengthL
       members = foldMap' (^. committeeMembersL) committee
    in all (<= currentEpoch + fromIntegral maxCommitteeTerm) members